+++ devel/realtest/casemap-stringx.R
@@ -1,2 +1,0 @@
-E(toupper('i', locale='en_US'), "I")
-E(toupper('i', locale='tr_TR'), "\u0130")