+++ tests/casemap-all.R
@@ -1,3 +1,6 @@
+if (Sys.getenv("STRINGX_DO_NOT_LOAD") != "1") library("stringx")
+library("realtest")
+
 E(tolower(character(0)), character(0))
 E(tolower(c("a", NA_character_, "c")), c("a", NA_character_, "c"))
 E(toupper(character(0)), character(0))
@@ -27,17 +30,17 @@ E(
     bad=c("gro\u00DF", "gross", "gross")
 )
 
-E(
-    toupper(stringx::letters_greek),
-    stringx::LETTERS_GREEK,
-    bad=stringx::letters_greek
-)
 
-E(
-    tolower(stringx::LETTERS_GREEK),
-    stringx::letters_greek,
-    bad=stringx::LETTERS_GREEK
-)
+E(toupper(stringx::letters_greek), stringx::LETTERS_GREEK,   bad=stringx::letters_greek)
+E(tolower(stringx::LETTERS_GREEK), stringx::letters_greek,   bad=stringx::LETTERS_GREEK)
+E(toupper(stringx::letters_bf),    stringx::LETTERS_BF,      bad=stringx::letters_bf)
+E(tolower(stringx::LETTERS_BF),    stringx::letters_bf,      bad=stringx::LETTERS_BF)
+E(toupper(stringx::letters_bb),    stringx::LETTERS_BB,      bad=stringx::letters_bb)
+E(tolower(stringx::LETTERS_BB),    stringx::letters_bb,      bad=stringx::LETTERS_BB)
+E(toupper(stringx::letters_cal),   stringx::LETTERS_CAL,     bad=stringx::letters_cal)
+E(tolower(stringx::LETTERS_CAL),   stringx::letters_cal,     bad=stringx::LETTERS_CAL)
+E(toupper(stringx::letters_frak),  stringx::LETTERS_FRAK,    bad=stringx::letters_frak)
+E(tolower(stringx::LETTERS_FRAK),  stringx::letters_frak,    bad=stringx::LETTERS_FRAK)